Tarefa2_2021li1g021

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

A função constroiMapa contrói um mapa

Exemplos de utilização

>>> constroiMapa [(Porta, (0, 3)),(Bloco, (0, 4)),(Bloco, (1, 4)),(Bloco, (2, 4)),(Bloco, (3, 4)),(Bloco, (4, 4)),(Caixa, (4, 3)),(Bloco, (5, 4)),(Bloco, (6, 4)),(Bloco, (6, 3)),(Bloco, (6, 2)),(Bloco, (6, 1))]

[[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio]
,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Bloco]
,[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Bloco]
,[Porta,Vazio,Vazio,Vazio,Caixa,Vazio,Bloco]
,[Bloco,Bloco,Bloco,Bloco,Bloco,Bloco,Bloco]]

constroiMapaAUX :: [(Peca, Coordenadas)] -> Mapa -> Int -> Int -> [Peca] -> Mapa

A função constroiMapaAUX é uma auxiliar da função constroiMapa que pega numa lista ordenada e mete os seus elementos nos sítios certos de um mapa de Vazios com a mesma dimensão

quantosVaziosFaltam :: [Peca] -> [Peca] -> [Peca]

A função quantosVaziosFaltam dá uma lista de Vazios do tamanho necessário para juntá-la a uma certa lista de modo a que esta fique do mesmo tamanho da 2ª lista que recebe

Exemplos de utilização

>>> quantosVaziosFaltam [Porta] [Vazio,Vazio,Vazio,Vazio]

[Vazio,Vazio,Vazio]

>>> [Porta] ++ quantosVaziosFaltam [Porta] [Bloco,Porta,Caixa,Vazio,Bloco]

[Porta,Vazio,Vazio,Vazio,Vazio]

ordenaXeY :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função ordenaXeY ordena uma lista do tipo [(Peca, Coordenadas)] de acordo com as suas coordenadas

Exemplos de utilização

>>> ordenaXeY [(Porta,(2,5)), (Bloco,(3,2)), (Bloco,(2,3)), (Bloco,(1,2))]

[(Bloco,(1,2)),(Bloco,(3,2)),(Bloco,(2,3)),(Porta,(2,5))]

ordenaXeYaux :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função ordenaXeYaux é uma auxiliar da ordenaXeY

ordenaY :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função ordenaY ordena uma lista relativamente ao y da coordenada

Exemplos de utilização

>>> ordenaY [(Porta,(2,5)), (Bloco,(1,2)), (Bloco,(2,3)), (Bloco,(3,2))]

[(Bloco,(3,2)),(Bloco,(1,2)),(Bloco,(2,3)),(Porta,(2,5))]

auxOrdenaY :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função auxOrdenaY é uma auxiliar da função ordenaY

ordenaX :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função ordenaX ordena uma lista relativamente ao x da coordenada

Exemplos de utilização

>>> ordenaX [(Porta,(2,5)), (Bloco,(1,2)), (Bloco,(2,3)), (Bloco,(3,2))]

[(Bloco,(1,2)),(Bloco,(2,3)),(Porta,(2,5)),(Bloco,(3,2))]

auxOrdenaX :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

A função auxOrdenaX é uma auxiliar da função ordenaX

criarMapaVazio :: Coordenadas -> Mapa

A função criarMapaVazio pega numa coordenada e cria um mapa de Vazios com essas dimensões

Exemplos de utilização

>>> criarMapaVazio (2,3)

[[Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio]]

criarMapaVazioAUX1 :: Coordenadas -> Int -> [Peca] -> [[Peca]]

A função criarMapaVazioAUX1 é uma auxiliar da função criarMapaVazio que pega nas listas dadas pela função criarMapaVazioAUX2 e cria uma lista dessas listas

criarMapaVazioAUX2 :: Coordenadas -> Coordenadas -> [Peca] -> [Peca]

A função criarMapaVazioAUX2 é uma auxiliar da função criarMapaVazio que dá uma linha do mapa de cada vez

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

A função desconstroiMapa recebe um mapa e transforma-o na forma [(Peca, Coordenadas)], ocultando os Vazios

Exemplos de utilização

>>> desconstroiMapa [[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Bloco],[Vazio,Vazio,Vazio,Vazio,Vazio,Vazio,Bloco],[Porta,Vazio,Vazio,Vazio,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco,Bloco,Bloco]]

[(Bloco,(6,1)),(Bloco,(6,2)),(Porta,(0,3)),(Caixa,(4,3)),(Bloco,(6,3)),(Bloco,(0,4)),(Bloco,(1,4)),(Bloco,(2,4)),(Bloco,(3,4)),(Bloco,(4,4)),(Bloco,(5,4)),(Bloco,(6,4))]

desconstroiMapaAUX :: Mapa -> Coordenadas -> [(Peca, Coordenadas)]

A função desconstroiMapaAUX é uma auxiliar da função desconstroiMapa