Tarefa4_2021li1g021

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

moveJogador :: Jogo -> Movimento -> Jogo

A função moveJogador chama uma função que aplica o movimento ao jogo, seja ele qual for

interageCaixa :: Mapa -> Jogador -> Jogo

A função interageCaixa chama uma função que pega numa caixa ou que pousa uma caixa dependendo da situação

pegaCaixa :: Mapa -> Jogador -> Jogo

A função pegaCaixa faz o jogador pegar na caixa que está na sua frente, se possível

não pega quando não tem uma caixa à frente

    |dir == Este && (elem (Caixa,(x+1,y)) (desconstroiMapa mapa) == False) = Jogo mapa (Jogador (x,y) dir bool)
    |dir == Oeste && (elem (Caixa,(x-1,y)) (desconstroiMapa mapa) == False) = Jogo mapa (Jogador (x,y) dir bool)

não pega quando a caixa tem um obstáculo em cima

    |dir == Este && (elem (Bloco,(x+1,y-1)) (desconstroiMapa mapa) || elem (Caixa,(x+1,y-1)) (desconstroiMapa mapa)) = Jogo mapa (Jogador (x,y) dir bool)
    |dir == Oeste && (elem (Bloco,(x-1,y-1)) (desconstroiMapa mapa) || elem (Caixa,(x-1,y-1)) (desconstroiMapa mapa)) = Jogo mapa (Jogador (x,y) dir bool) 

não pega quando o jogador tem obstáculo em cima

    |elem (Bloco,(x,y-1)) (desconstroiMapa mapa) = Jogo mapa (Jogador (x,y) dir bool)

tiraCaixaMapa :: [(Peca, Coordenadas)] -> Coordenadas -> [(Peca, Coordenadas)] -> Mapa

a função tiraCaixaMapa retira uma determinada caixa do mapa

Exemplos de utilização

>>> tiraCaixaMapa [(Porta,(2,5)), (Bloco,(3,2)), (Caixa,(2,3)), (Caixa,(1,2))] (2,3) []

[[Vazio,Vazio,Vazio,Vazio],[Vazio,Vazio,Vazio,Vazio],[Vazio,Caixa,Vazio,Bloco],[Vazio,Vazio,Vazio,Vazio],[Vazio,Vazio,Porta,Vazio]]

pousaCaixa :: Mapa -> Jogador -> Jogo

A função pousaCaixa faz o jogador pousar a caixa que carrega, se for possível

quando não pode pousar a caixa

    |dir == Este && (elem (Caixa,(x+1,y-1))(desconstroiMapa mapa) || elem (Bloco,(x+1,y-1))(desconstroiMapa mapa))  = Jogo mapa (Jogador (x,y) dir bool) 
    |dir == Oeste && (elem (Caixa,(x-1,y-1))(desconstroiMapa mapa) || elem (Bloco,(x-1,y-1))(desconstroiMapa mapa))  = Jogo mapa (Jogador (x,y) dir bool)

quando empilha a caixa em cima de outra coisa

    |dir == Este && (elem (Caixa,(x+1,y))(desconstroiMapa mapa) || elem (Bloco,(x+1,y))(desconstroiMapa mapa)) && (elem (Porta,(x+1,y-1)) (desconstroiMapa mapa) == False) = Jogo (empilhaCaixaDir (desconstroiMapa mapa) (x,y)) (Jogador (x,y) dir False)
    |dir == Oeste && (elem (Caixa,(x-1,y))(desconstroiMapa mapa) || elem (Bloco,(x-1,y))(desconstroiMapa mapa)) && (elem (Porta,(x-1,y-1)) (desconstroiMapa mapa) == False) = Jogo (empilhaCaixaEsq (desconstroiMapa mapa) (x,y)) (Jogador (x,y) dir False)

quando pousa a caixa (à direita ou à esquerda)

    |dir == Este && (elem (Porta,(x+1,y)) (desconstroiMapa mapa) == False) = Jogo (pousaDir (desconstroiMapa mapa) (x,y)) (Jogador (x,y) dir False)
    |dir == Oeste && (elem (Porta,(x-1,y)) (desconstroiMapa mapa) == False) = Jogo (pousaEsq (desconstroiMapa mapa) (x,y)) (Jogador (x,y) dir False)

pousaDir :: [(Peca, Coordenadas)] -> Coordenadas -> Mapa

a função pousaDir acrescenta uma caixa no mapa na coordenada à direita do jogador

Exemplos de utilização

>>> pousaDir [(Porta,(2,5)), (Bloco,(3,2)), (Caixa,(2,3)), (Caixa,(1,2))] (3,3)

[[Vazio,Vazio,Vazio,Vazio,Vazio]
,[Vazio,Vazio,Vazio,Vazio,Vazio]
,[Vazio,Caixa,Vazio,Bloco,Vazio]
,[Vazio,Vazio,Caixa,Vazio,Caixa]
,[Vazio,Vazio,Porta,Vazio,Vazio]]

pousaEsq :: [(Peca, Coordenadas)] -> Coordenadas -> Mapa

a função pousaEsq acrescenta uma caixa no mapa na coordenada à esquerda do jogador

Exemplos de utilização

>>> pousaEsq [(Porta,(2,5)), (Bloco,(3,2)), (Caixa,(2,3)), (Caixa,(1,2))] (1,1)

[[Vazio,Vazio,Vazio,Vazio]
,[Caixa,Vazio,Vazio,Vazio]
,[Vazio,Caixa,Vazio,Bloco]
,[Vazio,Vazio,Caixa,Vazio]
,[Vazio,Vazio,Porta,Vazio]]

quedaCaixaDir :: [(Peca, Coordenadas)] -> Coordenadas -> Mapa

a função quedaCaixaDir dá o mapa depois do jogador pousar a caixa à sua direita e depois desta cair, já que ocorre sempre que não existe chão à esquerda do jogador

quedaCaixaEsq :: [(Peca, Coordenadas)] -> Coordenadas -> Mapa

a função quedaCaixaEsq dá o mapa depois do jogador pousar a caixa à sua esquerda e depois desta cair, já que ocorre sempre que não existe chão à esquerda do jogador

empilhaCaixaDir :: [(Peca, Coordenadas)] -> Coordenadas -> Mapa

a função empilhaCaixaDir dá o mapa depois do jogador empilhar a caixa em cima de algo à direita

empilhaCaixaEsq :: [(Peca, Coordenadas)] -> Coordenadas -> Mapa

a função empilhaCaixaEsq dá o mapa depois do jogador empilhar a caixa em cima de algo à esquerda

trepar :: Mapa -> Jogador -> Jogo

a função trepar retorna o jogo depois do jogador trepar algo(se possível)

quando o elemento diretamente ao lado do jogador n existe

    |dir == Este && elem (Bloco,(x+1,y)) (desconstroiMapa mapa) == False && elem (Caixa,(x+1,y)) (desconstroiMapa mapa) == False = Jogo mapa (Jogador (x,y) dir bool)  
    |dir == Oeste && elem (Bloco,(x-1,y)) (desconstroiMapa mapa) == False && elem (Caixa,(x-1,y)) (desconstroiMapa mapa) == False = Jogo mapa (Jogador (x,y) dir bool)

quando há 2 blocos ou caixas empilhados diretamente ao lado do jogador

    |dir == Este && (elem (Bloco,(x+1,y-1)) (desconstroiMapa mapa)) || elem (Caixa,(x+1,y-1)) (desconstroiMapa mapa) = Jogo mapa (Jogador (x,y) dir bool)
    |dir == Oeste && (elem (Bloco,(x-1,y-1)) (desconstroiMapa mapa)) || elem (Caixa,(x-1,y-1)) (desconstroiMapa mapa) = Jogo mapa (Jogador (x,y) dir bool)

quando o jogador estiver a carregar uma caixa e não houver 2 espaços por cima do bloco que o jogador quer trepar

    |dir == Este && bool == True && elem (Bloco,(x+1,y-2)) (desconstroiMapa mapa) = Jogo mapa (Jogador (x,y) dir bool)
    |dir == Oeste && bool == True && elem (Bloco,(x-1,y-2)) (desconstroiMapa mapa) = Jogo mapa (Jogador (x,y) dir bool)

quando o jogador trepa normalmente

    |dir == Este = Jogo mapa (Jogador (x+1,y-1) Este bool)
    |dir == Oeste = Jogo mapa (Jogador (x-1,y-1) Oeste bool)

andarDir :: Mapa -> Jogador -> Jogo

a função andarDir faz o jogador virar-se para a direita e andar uma vez, se for possível

andarEsq :: Mapa -> Jogador -> Jogo

a função andarEsq faz o jogador virar-se para a esquerda e andar uma vez, se for possível

ondeVaiCair :: Int -> Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> Int

a função ondeVaiCair dá o y da coordenada onde algo (jogador ou caixa) vai aterrar

moveJogadorAUX :: Jogo -> Movimento -> [Movimento] -> Jogo

a função moveJogadorAUX é uma auxiliar da função correrMovimentos que faz com que um jogador se mova várias vezes (fazendo com que a função moveJogador execute várias vezes)

correrMovimentos :: Jogo -> [Movimento] -> Jogo

a função correrMovimentos faz com que sejam executados vários movimentos a um jogo